requirements.txt
guilded.py
colorama